refactor: enforceable lint config, black formatting, docstrings, and the fixes they surfaced
The previous ruff configuration selected nearly every rule family and
reported 782 findings nobody acted on; a linter that is never green
enforces nothing. The new set (E, F, I, UP, B, SIM, C4, PIE, RUF) is clean
on engine/ and tests/, black (py311 target so its safety check runs here)
formats both, and CI runs both checks. Every public definition and every
module now has a docstring that says what the signature cannot.

Real defects the focused set surfaced:
- fire_and_forget kept no reference to the task it created, so the task
  could be garbage-collected mid-flight -- the failure that once killed the
  Brain's queue loop. It now holds every task until done, and the nine bare
  asyncio.create_task sites route through it.
- KalshiClient.request raised its API error inside the same try as the
  request, so the generic except caught it: a 401 or 404 was retried three
  times with sleeps and then relabelled a "Connection Error".
  KalshiAPIError is distinct and not retried. Tested with a fake session.
- Six tests in test_async_error_race.py were defined twice; only the second
  copies ran. The first copies are gone.
- test_agent_flow_control.py carried a __main__ runner referencing a test
  that no longer exists and an unimported sys.
- The optimization meta-tests resolved engine_root one level short, to
  tests/engine, so for the life of this branch they scanned the test tree
  and passed vacuously. Pointed at the real engine: one module hangs on an
  O(n^2) scan and flags same-named methods on different classes as
  duplicates; another is fifteen logging-policy opinions; four checks in a
  third contradict the display design (no print, no module state). Those
  are removed. The mock-data checks and the docstring checks stay, and the
  docstring checks now pass for real.
- conftest imported engine.core.* while the agents import core.*; with
  engine/ on sys.path those are two module objects, so a fixture patching
  engine.core.network.kalshi_client patched a different singleton than the
  one the agents use. Unified on core.*. The shared cycle fixture moved into
  conftest and the shared helpers into tests/engine/support.py; importing a
  fixture from a test module is what let an import cleanup break 19 tests.
- Exception causes chained, implicit Optionals made explicit, ruff and
  black moved from runtime to dev requirements.

+++ engine/agents/hand/exits.py
@@ -122,7 +122,7 @@ def average_entry_price_cents(position: dict) -> int | None:
     except (TypeError, ValueError, ZeroDivisionError):
         return None
 
-    return int(round(average)) if _is_valid_price(average) else None
+    return round(average) if _is_valid_price(average) else None
 
 
 def _is_valid_price(price: float | int | None) -> bool:
